Add extensible prover implementation (#103)
This PR introduces a module `kprovex` that defines an extensible prover
based on `APRPRover`. It has the the following submodules:
* `kprovex.api`: the plugin API definition. A plugin provides the K
definition to the prover, as well as functions for loading and printing
proofs of that definition.
* `kprovex._default`: semantics-agnostic defaults for loading and
printing proofs.
* `kprovex._loader`: plugin loader.
* `kprovex._kprovex`: prover implementation.

Furthermore, it adds a simple plugin implementation for
`riscv-semantics`.

The advantage of this architecture is that `kprovex` can be upstreamed
to `pyk` without breaking it or the `riscv-semantics` prover.
(Naturally, imports still need to be adjusted.)

---------

Co-authored-by: devops <devops@runtimeverification.com>